Renamed existing component to kern dropdown
