+++ src/app/base/base.module.ts
@@ -16,7 +16,7 @@ import { SidebarPmComponent } from './components/sidebar-pm/sidebar-pm.component
 import { DropdownDirective } from './directives/dropdown.directive';
 import { PercentRoundPipe } from './pipes/decimal-round.pipe';
 import { HeaderComponent } from './components/header/header.component';
-import { DropdownItComponent } from './components/dropdown-it/dropdown-it.component';
+import { DropdownComponent } from './components/dropdown/dropdown.component';
 
 @NgModule({
   declarations: [
@@ -28,7 +28,7 @@ import { DropdownItComponent } from './components/dropdown-it/dropdown-it.compon
     DropdownDirective,
     PercentRoundPipe,
     HeaderComponent,
-    DropdownItComponent
+    DropdownComponent
   ],
   imports: [CommonModule, AppRoutingModule, HighlightModule],
   exports: [
@@ -50,7 +50,7 @@ import { DropdownItComponent } from './components/dropdown-it/dropdown-it.compon
     DropdownDirective,
     PercentRoundPipe,
     HeaderComponent,
-    DropdownItComponent
+    DropdownComponent
   ],
   providers: [
     {
