+++ src/app/base/components/dropdown/dropdown-helper.ts
