+++ src/app/base/components/dropdown/dropdown.component.scss
