+++ src/app/base/components/header/header.component.html
@@ -60,12 +60,12 @@
             </div>
             <div class="flex justify-center overflow-visible">
                 <div *ngIf="user" class=" mr-4">
-                        <kern-dropdown-it [dropdownOptions]="{
+                        <kern-dropdown [dropdownOptions]="{
                             optionArray: showConfigSettings ? ['Account Settings', 'Change config', 'Logout'] : ['Account Settings', 'Logout'],
                             buttonVersion: 'userIcon',
                             avatarUri: avatarUri
                         }" (optionClicked)="executeOption($event)"
-                        ></kern-dropdown-it>
+                        ></kern-dropdown>
                 </div>
             </div>
         </div>
