+++ src/app/weak-supervision/components/weak-supervision/weak-supervision.component.html
@@ -27,13 +27,13 @@
                         </button>
                     </div>
                     <ng-template [ngIf]="labelingTasks && labelingTasks.length > 0" [ngIfElse]="noTasks">
-                        <kern-dropdown-it class="mr-3" [dropdownOptions]="{
+                        <kern-dropdown class="mr-3" [dropdownOptions]="{
                             optionArray: ['Labeling function', 'Active learning', 'Zero-shot'],
                             optionIcons: ['labeling-function','active-learning','zero-shot'],
                             buttonCaption: 'New heuristic',
                             buttonTooltip: 'Heuristics are noisy label signals'
                         }" (optionClicked)="executeOption($event)">
-                        </kern-dropdown-it>
+                        </kern-dropdown>
                     </ng-template>
                     <ng-template #noTasks>
                         <div data-tip="At least one labeling task is needed to create Heuristics"
@@ -53,14 +53,14 @@
                     </ng-template>
                     
                     <ng-template [ngIf]="informationSources.length > 0" [ngIfElse]="noHeuristics">
-                        <kern-dropdown-it class="mr-3" [dropdownOptions]="{
+                        <kern-dropdown class="mr-3" [dropdownOptions]="{
                             optionArray:['Select all','Deselect all', 'Run selected', 'Delete selected'],
                             optionIcons:['select-all','deselect-all', 'run', 'delete-selected'],
                             buttonTooltip:'Multi-heuristic action',
                             buttonCaption: 'Actions',
                             isOptionDisabled: [false, false, this.selectedInformationSources.length == 0, this.selectedInformationSources.length == 0]
                         }" (optionClicked)="executeOption($event)">
-                        </kern-dropdown-it>
+                        </kern-dropdown>
                     </ng-template>
                     <ng-template #noHeuristics>
                         <div data-tip="At least one heuristic is needed to enable actions"
